+++ nym-vpn-android/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0]
+
 ### Added
 - Add subscription checks after login (https://github.com/nymtech/nym-vpn-client/pull/4683)
 - Expose account summary for handling Account state (https://github.com/nymtech/nym-vpn-client/pull/4683)
@@ -19,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix No subscription error handling on Main screen (https://github.com/nymtech/nym-vpn-client/pull/4685)
 - Error handling during login (https://github.com/nymtech/nym-vpn-client/pull/4708)
+- Fix Boot receiver (https://github.com/nymtech/nym-vpn-client/pull/4730)
 
 ## [2.9.0] - 18.02.2026